2.1.0: refactor; add progress report
This started off as an attempt to use lookup tables to reduce the size
of the script. It worked: the code of `journal_entry` was 70% smaller,
and overall it was 45% smaller even if you insisted on counting the
lookup tables.

I could have merged there, but I didn't.

Instead, I added a progress report so we didn't need as many secret
ninja commands. It was tidier if I parked it off to the side, but that
meant I needed a package style layout and a new update mechanism. And...
well, around then it got away from me. Sorry about that, truckers.

Entarius Fusion's new route makes the progress update nice and fast.