Turn on progress tracking by default #23
Required rebuilding the preferences system. There's no way with the EDMC
config system to default an integer to 1. So, we layer our own on top.
Ours behaves just like a Python `dict`, EXCEPT:

* All values are stored as JSON

* A value of `None` is treated as if the value is absent

* We have to guess the full list of `keys` because EDMC provides no way
  of discovering them

We use Python's inbuilt `UserDict.DictMixin` to keep the code small and
the behaviour compatible. All your favourite Python `dict` methods are
present, including `get` and `setdefault`.

We also prefix our EDMC config keys with `HuttonHelper` so there's less
chance of a conflict.

To set the default value of `PrefName` in your plugin's `__init__`,
`plugin_app`, or `plugin_prefs`:

    self.helper.prefs.setdefault('PrefName', default_value)

To get the current value in a method called later enough that you KNOW
the value must be present:

    value = self.helper.prefs['PrefName']

To set it:

    self.helper.prefs['PrefName'] = value