Auto resize
I tried using this SelfWrappingHyperlinkLabel in one of my own projects and found that it wasn't resizing properly. It was always wrapping too short. This might only be apparent if you are using lots of plugins.  So I modified to bind to the parent frame <Configure> event and now it seems to work fine.