Panic #58
It's somewhat pleasant to have **DON'T PANIC!** on the front cover if we
have nothing else to tell the trucker apart from the news. We also keep
it up there for the first few seconds after launch because clicking on
controls won't work if any of the plugins are talking to their servers,
and startup is a busy time.

That said, some truckers would prefer we take up less room on their EDMC
console, and fair enough.

Under EDMC > Settings > Hutton Helper > Big Friendly Letters Control,
check the box marked PANIC to have the message disappear once the Helper
has given all the EDMC plugins time to talk to their servers.