Mitigate problem with explocredit #62
If the third attempt to get the exploration credit fails, the plugin
will:

* Stop trying so the Helper doesn't hit the server once per event
  until ~~the heat death of the universe~~ they stop playing

* Display `Disabled (#62)` instead of the credit count

Truckers without the feature enabled in Settings won't see a thing.
Those that do have it enabled, will hopefully be motivated enough to
let us know it's not working while we figure out why they're not
getting the JSON they want.